+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fix
 
+- **base/flattener.py**: fix problem with flatten method, it should be abstract
 - **commons.py**: a fix was made for a problem with "| None"
 
 ## v1.1.1 (2024-09-02)